DEV: Move theme hbs transforms into dedicated file
Extracted from #33103
