+++ app/assets/javascripts/theme-transpiler/transpiler.js
@@ -12,40 +12,12 @@ import { minify as terserMinify } from "terser";
 import { WidgetHbsCompiler } from "discourse-widget-hbs/lib/widget-hbs-compiler";
 import { browsers } from "../discourse/config/targets";
 import { Preprocessor } from "./content-tag";
+import buildEmberTemplateManipulatorPlugin from "./theme-hbs-ast-transforms";
 
 const thisFallbackPlugin = EmberThisFallback._buildPlugin({
   enableLogging: false,
   isTheme: true,
 }).plugin;
-
-function manipulateAstNodeForTheme(node, themeId) {
-  // Magically add theme id as the first param for each of these helpers)
-  if (
-    node.path.parts &&
-    ["theme-i18n", "theme-prefix", "theme-setting"].includes(node.path.parts[0])
-  ) {
-    if (node.params.length === 1) {
-      node.params.unshift({
-        type: "NumberLiteral",
-        value: themeId,
-        original: themeId,
-        loc: { start: {}, end: {} },
-      });
-    }
-  }
-}
-
-function buildEmberTemplateManipulatorPlugin(themeId) {
-  return function () {
-    return {
-      name: "theme-template-manipulator",
-      visitor: {
-        SubExpression: (node) => manipulateAstNodeForTheme(node, themeId),
-        MustacheStatement: (node) => manipulateAstNodeForTheme(node, themeId),
-      },
-    };
-  };
-}
 
 function buildTemplateCompilerBabelPlugins({ extension, themeId }) {
   const compiler = { precompile };
